+++ CHANGES.md
@@ -5,6 +5,7 @@ Version 1.11.0
 --------------
 
 * Add support for producing reports in HTML format.
+* Add support for specifying the title of the report.
 * Fix incorrect count of articles in ZIP file comments.